+++ src/Symfony/Component/HttpKernel/Tests/HttpKernelBrowserTest.php
@@ -179,4 +179,15 @@ class HttpKernelBrowserTest extends TestCase
 
         unlink($source);
     }
+
+    public function testAcceptHeaderNotSet()
+    {
+        $client = new HttpKernelBrowser(new TestHttpKernel());
+
+        $client->request('GET', '/');
+        $this->assertFalse($client->getRequest()->headers->has('Accept'));
+
+        $client->request('GET', '/', [], [], ['HTTP_ACCEPT' => 'application/ld+json']);
+        $this->assertSame('application/ld+json', $client->getRequest()->headers->get('Accept'));
+    }
 }