+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
  * Allowed adding attributes on controller arguments that will be passed to argument resolvers.
  * kernels implementing the `ExtensionInterface` will now be auto-registered to the container
  * added parameter `kernel.runtime_environment`, defined as `%env(default:kernel.environment:APP_RUNTIME_ENV)%`
+ * do not set a default `Accept` HTTP header when using `HttpKernelBrowser`
 
 5.1.0
 -----